+++ modules/local/diann/preliminary_analysis/main.nf
@@ -106,9 +106,12 @@ process PRELIMINARY_ANALYSIS {
             --no-prot-inf \\
             ${diann_dda_flag} \\
             \${mod_flags} \\
-            $args
+            $args \\
+            2>&1 | tee ${ms_file.baseName}_diann.log
 
-    cp report.log.txt ${ms_file.baseName}_diann.log
+    if [ -f report.log.txt ]; then
+        cp report.log.txt ${ms_file.baseName}_diann.log
+    fi
 
     cat <<-END_VERSIONS > versions.yml
     "${task.process}":
